[BE] fix: fix running domain
S12P31A502-140

- fix running start
- fix running end
- fix create running point

+++ backend/src/main/java/com/ssafy/yoittang/running/domain/dto/request/ChallengeRunningCreateRequest.java
@@ -10,10 +10,10 @@ public record ChallengeRunningCreateRequest(
         Long courseId,
 
         @NotNull
-        Long lat,
+        Double lat,
 
         @NotNull
-        Long lng,
+        Double lng,
 
         @NotNull
         LocalDateTime currentTime
