+++ backend/src/main/java/com/ssafy/yoittang/runningpoint/domain/dto/request/GeoPoint.java
@@ -5,8 +5,8 @@ import jakarta.validation.constraints.NotNull;
 public record GeoPoint(
 
         @NotNull
-        Long lat,
+        Double lat,
 
         @NotNull
-        Long lng
+        Double lng
 ) { }
